update: patch for zor api setup in test
